Simulated Ebola outbreak

This simulated outbreak of Ebola Virus Disease matches some key properties of the West African Ebola outbreak of 2014-2015. Specifically, care was taken to use realistic delays (incubation period, serial interval, time to hospitalisation, etc.) and reproduction number (see references).

ebola.sim

Format

An object of class list of length 2.

Details

This dataset is used for teaching purposes during Imperial College's short course on infectious disease modelling. The exercise aims to simulate the response to an Ebola outbreak taking place in a single large city, and evaluate the impact of an intervention (increased bed capacity).

Note that to ensure realism, some errors have been introduced in this dataset. These can be identified as negative incubation periods (delay from infection to onset of symptoms). See example for a simple way to identify these cases. The dataset sim.ebola.clean in the 'outbreaks' package is the same dataset, only dates of infection and onset have been set to 'NA'.

References

WHO Ebola Response Team, 2014. Ebola virus disease in West Africa--the first 9 months of the epidemic and forward projections. The New England journal of medicine 371:1481-1495.

WHO Ebola Response Team, 2015. West African Ebola epidemic after one year -- slowing but not yet under control. The New England journal of medicine 372:584-587.

See also

The 'outbreaks' package, from which this dataset is taken.

Examples

## check content
names(ebola.sim)

#> [1] "linelist" "contacts"

## check line lists
head(ebola.sim$linelist)

#>   case.id generation date.of.infection date.of.onset date.of.hospitalisation
#> 1  d1fafd          0              <NA>    2014-04-07              2014-04-17
#> 2  53371b          1        2014-04-09    2014-04-15              2014-04-20
#> 3  f5c3d8          1        2014-04-18    2014-04-21              2014-04-25
#> 4  6c286a          2              <NA>    2014-04-27              2014-04-27
#> 5  0f58c4          2        2014-04-22    2014-04-26              2014-04-29
#> 6  49731d          0        2014-03-19    2014-04-25              2014-05-02
#>   date.of.outcome outcome gender           hospital
#> 1      2014-04-19    <NA>      f  Military Hospital
#> 2            <NA>    <NA>      m Connaught Hospital
#> 3      2014-04-30 Recover      f              other
#> 4      2014-05-07   Death      f               <NA>
#> 5      2014-05-17 Recover      f              other
#> 6      2014-05-07    <NA>      f               <NA>

## check contacts
head(ebola.sim$contacts)

#>    infector case.id  source
#> 2    d1fafd  53371b   other
#> 3    cac51e  f5c3d8 funeral
#> 5    f5c3d8  0f58c4   other
#> 8    0f58c4  881bd4   other
#> 11   8508df  40ae5f   other
#> 12   127d83  f547d6 funeral