Find node degree for epi_contacts objects

This function computes the number of contacts per cases in a epi_contacts dataset. Whenever contacts are directed, the argument 'type' can be used to specify which kind of contact should be considered: 'in' (towards the case), 'out' (from the case), or 'both'.

get_degree(x, type = c("in", "out", "both"), only_linelist = FALSE)

Arguments

x

an epi_contacts object

type

the type of degree to be computed (see description); if contacts are not directed, this will be forced to 'both'

only_linelist

a logical indicating if cases whose degree is computed should be from the linelist exclusively

Examples

x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)
x

#> 
#> /// Epidemiological Contacts //
#> 
#>   // class: epi_contacts
#>   // 5,888 cases in linelist; 3,800 contacts;  directed 
#> 
#>   // linelist
#> 
#> # A tibble: 5,888 × 9
#>        id generation date.of.infection date.of.onset date.of.hospitalisation
#> *   <chr>      <int>            <date>        <date>                  <date>
#> 1  d1fafd          0              <NA>    2014-04-07              2014-04-17
#> 2  53371b          1        2014-04-09    2014-04-15              2014-04-20
#> 3  f5c3d8          1        2014-04-18    2014-04-21              2014-04-25
#> 4  6c286a          2              <NA>    2014-04-27              2014-04-27
#> 5  0f58c4          2        2014-04-22    2014-04-26              2014-04-29
#> 6  49731d          0        2014-03-19    2014-04-25              2014-05-02
#> 7  f9149b          3              <NA>    2014-05-03              2014-05-04
#> 8  881bd4          3        2014-04-26    2014-05-01              2014-05-05
#> 9  e66fa4          2              <NA>    2014-04-21              2014-05-06
#> 10 20b688          3              <NA>    2014-05-05              2014-05-06
#> # ... with 5,878 more rows, and 4 more variables: date.of.outcome <date>,
#> #   outcome <fctr>, gender <fctr>, hospital <fctr>
#> 
#>   // contacts
#> 
#> # A tibble: 3,800 × 3
#>      from     to  source
#> *   <chr>  <chr>  <fctr>
#> 1  d1fafd 53371b   other
#> 2  cac51e f5c3d8 funeral
#> 3  f5c3d8 0f58c4   other
#> 4  0f58c4 881bd4   other
#> 5  8508df 40ae5f   other
#> 6  127d83 f547d6 funeral
#> 7  f5c3d8 d58402   other
#> 8  20b688 d8a13d   other
#> 9  2ae019 a3c8b8   other
#> 10 20b688 974bc1   other
#> # ... with 3,790 more rows
#> 

deg_in <- get_degree(x)
deg_out <- get_degree(x, "out")
deg_both <- get_degree(x, "both")
head(deg_in)

#> d1fafd cac51e f5c3d8 0f58c4 8508df 127d83 
#>      0      0      1      1      0      0 

table(deg_in)

#> deg_in
#>    0    1 
#> 1711 3800 

head(deg_out)

#> d1fafd cac51e f5c3d8 0f58c4 8508df 127d83 
#>      1      1      2      1      3      1 

table(deg_out)

#> deg_out
#>    0    1    2    3    4    5    6 
#> 2814 1867  606  183   34    6    1 

head(deg_both)

#> d1fafd cac51e f5c3d8 0f58c4 8508df 127d83 
#>      1      1      3      2      3      1 

table(deg_both)

#> deg_both
#>    1    2    3    4    5    6    7 
#> 3982 1093  341   71   20    3    1