2015 Middle East respiratory syndrome outbreak in South Korea

These datasets correspond to the initial information collected by the Epidemic Intelligence group at European Centre for Disease Prevention and Control (ECDC) during the first weeks of the outbreak of Middle East respiratory syndrome (MERS-CoV) outbreak (South Korea) in 2015. The data were used to follow the daily evolution of this outbreak using public information available. Information is compiled in two datasets:

mers_kor_14

Format

An object of class list of length 2.

Details

- a line listing of MERS-CoV cases with a set of variables describing the patient's characteristics as follow: id (unique identifier), status (MERCoV status), age, age_class (age using 10-years groups), sex, place_infect (probable region of infection), reporting_ctry (country reporting case), loc_hosp (local hospital name where the case was hospitalized), dt_onset (date of onset of symptoms), dt_report (date of reporting), week_report (week number of date of reporting), dt_start_exp (date of first probable exposure to another MERCoV case), dt_end_exp (date of last probable exposure MERCoV case), dt_diag (date of MERCoV diagnosis) and dt_death (date of death).

Note that this dataset do not contain possible contact, therefore the network visualisation graphic will correspond to MERCoV transmission chains as the variable status corresponds to "case" only. (i.e. one node in the network represents one case).

- a database corresponding to the relationship between MERS-CoV cases as follow: from (unique identifier from the probable source patient), to (unique identifier from the secondary case), exposure (probable place of exposure) and diff_dt_onset (time in days between two successive cases). This second dataset is used to map the relationship between cases (i.e. edge in the transmission chain network). This dataset is used for teaching purposes only and intend to be use with the package "contacts" in order to illustrate network visualisation of transmission chains in epidemiology. This example do not represent nor the final outbreak investigation results neither the consolidated and complete description of the transmission chain.

References

More information on the intial stage of the outbreak in the following reference: Penttinen PM, Kaasik-Aaslav K, Friaux A, Donachie A, Sudre B, Amato-Gauci AJ, Memish ZA, Coulombier D. Taking stock of the first 133 MERS coronavirus cases globally--Is the epidemic changing? Euro Surveill. 2013 Sep 26;18(39). pii: 20596. PubMed PMID: 24094061.

Examples

## example using MERS outbreak in Korea, 2014
head(mers_kor_14[[1]])

#>     id age age_class sex        place_infect reporting_ctry
#> 1 SK_1  68     60-69   M         Middle East    South Korea
#> 2 SK_2  63     60-69   F Outside Middle East    South Korea
#> 3 SK_3  76     70-79   M Outside Middle East    South Korea
#> 4 SK_4  46     40-49   F Outside Middle East    South Korea
#> 5 SK_5  50     50-59   M Outside Middle East    South Korea
#> 6 SK_6  71     70-79   M Outside Middle East    South Korea
#>                                              loc_hosp   dt_onset  dt_report
#> 1 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-11 2015-05-19
#> 2 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-18 2015-05-20
#> 3 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-20 2015-05-20
#> 4 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-25 2015-05-26
#> 5                           365 Yeollin Clinic, Seoul 2015-05-25 2015-05-27
#> 6 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-24 2015-05-28
#>   week_report dt_start_exp dt_end_exp    dt_diag outcome   dt_death
#> 1     2014_21   2015-04-18 2015-05-04 2015-05-20   Alive       <NA>
#> 2     2014_21   2015-05-15 2015-05-20 2015-05-20   Alive       <NA>
#> 3     2014_21   2015-05-16 2015-05-16 2015-05-21    Dead 2015-06-04
#> 4     2014_22   2015-05-16 2015-05-20 2015-05-26   Alive       <NA>
#> 5     2014_22   2015-05-17 2015-05-17 2015-05-26   Alive       <NA>
#> 6     2014_22   2015-05-15 2015-05-17 2015-05-28    Dead 2015-06-01

head(mers_kor_14[[2]])

#>    from     to       exposure diff_dt_onset
#> 1 SK_14 SK_113 Emergency room            10
#> 2 SK_14 SK_116 Emergency room            13
#> 3 SK_14  SK_41 Emergency room            14
#> 4 SK_14 SK_112 Emergency room            14
#> 5 SK_14 SK_100 Emergency room            15
#> 6 SK_14 SK_114 Emergency room            15

x <- make_epi_contacts(linelist=mers_kor_14[[1]],
contacts=mers_kor_14[[2]], directed=TRUE)

## Not run: ------------------------------------
# plot(x)
# plot(x, group="place_infect")
# plot(x, group="loc_hosp", legend_max=20, annot=TRUE)
## ---------------------------------------------