Plot epicontacts objects

This function plots epicontacts objects using various approaches. The default method uses vis_epicontacts.

# S3 method for epicontacts
plot(
  x,
  node_color = "id",
  method = c("visNetwork", "graph3D", "ggplot"),
  x_axis = NULL,
  ...
)

Arguments

x
An epicontacts object

node_color
An index or character string indicating which field of the linelist should be used to color the nodes. If node color = 'R_i', the individual reproductive number for each case (i.e. number of outgoing infection/contacts) will be calculated and used to specify the node colour.

method
A character string indicating the plotting method to be used; available values are "visNetwork", "graph3D" and "ggplot". "ggplot" can only be used when the x_axis argument is specified.

x_axis
If a temporal axis is to be displayed, a character string indicating which field of the linelist data should be used to specify the x axis position (must be numeric or Date).

...
Further arguments passed to the plotting methods.

Details

This function is merely a wrapper for other plotting functions in the package, depending on the value of method:

visNetwork: calls the function vis_epicontacts when no x_axis argument is specified, and vis_temporal_interactive otherwise.

graph3D: calls the function graph3D.

ggplot: calls the function link{vis_temporal_static}.

See also

Author

Thibaut Jombart (thibautjombart@gmail.com)

Examples

if (require(outbreaks)) {
## example using MERS outbreak in Korea, 2014
head(mers_korea_2015[[1]])
head(mers_korea_2015[[2]])

x <- make_epicontacts(linelist = mers_korea_2015[[1]],
                      contacts = mers_korea_2015[[2]], directed=TRUE)

if (FALSE) {
plot(x)
plot(x, "place_infect")
plot(x, "loc_hosp", legend_max = 20, annot = TRUE)
plot(x, "place_infect", node_shape = "sex",
     shapes = c(M = "male", F = "female"))
plot(x, 4)
plot(x, 4, method = "graph3D")
}
}

#> Warning: Cycle(s) detected in the contact network: this may be unwanted