Subset an epi_contact object by factors

This function subsets an epicontacts object based on node, edge and/or cluster attributes. Specifying node attributes will return an epicontacts object containing only individuals with these given attributes in the linelist. Specifying edge attributes will return contacts with the attributes provided. Specifying cluster attributes will return clusters of connected cases, and can be defined by ids (returning clusters of cases connected to specified cases) or cluster sizes (returning cluster of a specific, minimum or maximum size).

# S3 method for epicontacts
subset(x, node.attribute = NULL,
  edge.attribute = NULL, cluster_id = NULL, cs = NULL, cs_min = NULL,
  cs_max = NULL, ...)

Arguments

x

an epi_contact object to be subsetted

node.attribute

a named list defining the node attribute name and node attribute value (as a single value or vector of values). Dates must be provided as a vector of date objects, defining the range of dates included in the subset

edge.attribute

a named list defining the edge attribute name and edge attribute value (as a single value or vector of values). Dates must be provided as a vector of date objects, defining the range of dates included in the subset

cluster_id

a character vector of case identifiers; the connected components attached to these cases will be retained in the output object.

cs

cluster size to be used for subsetting

cs_min

minimum cluster size for subsetting

cs_max

maximum cluster size for subsetting

...

further arguments passed on to other methods

Examples

if (require(outbreaks)) {
## build data
x <- make_epicontacts(ebola_sim$linelist, ebola_sim$contacts,
id="case.id", to="case.id", from="infector", directed=FALSE)

## subset based on node and edge attributes
x_subset <- subset(x, node.attribute=list("gender"="f"), edge.attribute=list("source"="funeral"))

## subset a cluster connected to a given id (can be a vector of ids as well)
id <- "cac51e"
x_subset <- subset(x, cluster_id=id)

## subset based on cluster size range
x_subset <- subset(x, cs_min = 12, cs_max = 15)

## subset based on single cluster size
x_subset <- subset(x, cs = 12)

## subset based on minimum cluster size
x_subset <- subset(x, cs_min = 10)

## subset based on maximum cluster size
x_subset <- subset(x, cs_max = 9)
}