Plot temporal networks using visNetwork

This function plots epicontacts objects with a temporal x-axis using the visNetwork package. The produced object is an htmlwidget which will need rendering within a web browser.

vis_temporal_interactive(
  x,
  x_axis = NULL,
  network_shape = c("branching", "rectangle"),
  root_order = x_axis,
  node_order = x_axis,
  reverse_root_order = FALSE,
  reverse_node_order = FALSE,
  rank_contact = x_axis,
  reverse_rank_contact = FALSE,
  unlinked_pos = c("bottom", "top", "middle"),
  edge_flex = FALSE,
  position_dodge = FALSE,
  parent_pos = c("middle", "top", "bottom"),
  custom_parent_pos = NULL,
  n_breaks = 5,
  axis_type = c("single", "double", "none"),
  igraph_type = NULL,
  timeline = NULL,
  tl_start_node_color = NULL,
  tl_end_node_color = NULL,
  tl_start_node_shape = NULL,
  tl_end_node_shape = NULL,
  tl_start_node_size = NULL,
  tl_end_node_size = NULL,
  tl_edge_color = NULL,
  tl_edge_width = NULL,
  tl_edge_linetype = NULL,
  tl_edge_label = NULL,
  ...
)

Arguments

x
An epicontacts object.

x_axis
A character string indicating which field of the linelist data should be used to specify the x axis position (must be numeric or Date).

network_shape
'branching' will create a branching transmission tree. 'rectangle' will create a rectangular shaped plot similar to a phylogeny that avoids overlapping edges.

root_order
A character string indicating which field of the linelist data is used to vertically order index cases of individual transmission chains (i.e. the 'roots' of the transmission trees). If root_order = 'subtree_size', index cases will be ordered by the size of the downstream transmission chains they generate.

node_order
A character string indicating which field of the linelist data is used to vertically order nodes in the transmission tree (i.e. the 'roots' of the transmission trees). If node_order = 'subtree_size', nodes will be ordered by the size of the downstream transmission chains they generate.

reverse_root_order
A logical indicating if the ordering of the roots should be reversed.

reverse_node_order
A logical indicating if the ordering of the nodes should be reversed.

rank_contact
If more than one incoming contact is provided for a given case, which attribute in the contact list should be used to rank the contacts and choose the top value. These contacts forms the 'scaffold' of the transmission tree and determine the y-position of each node. If a node attribute is provided, the pairwise difference in node attributes will be taken to rank the contacts.

reverse_rank_contact
Logical indicating if the contact ranking should be reversed in order.

unlinked_pos
A character string indicating where unlinked cases should be placed. Valid options are 'top', 'bottom' and 'middle', where 'middle' will place unlinked cases according to root_order.

edge_flex
A logical indicating if edges should bend when moved in visNetwork. If FALSE, edges will remain straight.

position_dodge
A logical indicating if two cases can occupy the same y coordinate or 'dodge' each other.

parent_pos
Specify the position of the parent node relative to its children. Can be one of 'middle', 'top' or 'bottom'.

custom_parent_pos
A function specifying the position of children nodes relative to their parent. This function must accept a single integer `x` as its only argument, specifying the number of children nodes. It must return a vector of length x, specifying the position of each child relative to the parent, where a x > 0 indicates above the parent, x < 0 indicates below the parent, and x = 0 indicates the same height as the parent.

n_breaks
The number of breaks on the x-axis timeline.

axis_type
Number of axes to be plotted (one of 'single', 'double', 'none')

igraph_type
Alternate network shapes provided by igraph. Must be one of 'rt' for Reingold-Tilford layout, 'sugiyama' for Sugiyama layout or 'fr' for Fruchterman-Reingold layout.

timeline
A dataframe of at least three columns: a column 'id' corresponding to the 'id' column in the linelist, a column 'start' of the same type as 'x_axis' indicating the start date of a timeline element, and a column 'end' of the same type as 'x_axis' indicating the end date of a timeline element.

tl_start_node_color
An index or character string indicating which field of the timeline should be used to color the start node of a timeline element.

tl_end_node_color
An index or character string indicating which field of the timeline should be used to color the end node of a timeline element.

tl_start_node_shape
An index or character string indicating which field of the timeline should be used to define the shape of the start node of a timeline element.

tl_end_node_shape
An index or character string indicating which field of the timeline should be used to define the shape of the end node of a timeline element.

tl_start_node_size
An index or character string indicating which field of the timeline should be used to define the size of the start node of a timeline element.

tl_end_node_size
An index or character string indicating which field of the timeline should be used to define the size of the end node of a timeline element.

tl_edge_color
An index or character string indicating which field of the timeline should be used to color the edge of a timeline element.

tl_edge_width
An index or character string indicating which field of the timeline should be used to define the width of the edge of a timeline element.

tl_edge_linetype
An index or character string indicating which field of the timeline should be used to define the linetype of the edge of a timeline element.

tl_edge_label
An index or character string indicating which field of the timeline should be used to label the edge of a timeline element.

...
Additional arguments specified in vis_epicontacts.

Value

The same output as visNetwork.

See also

Author

Finlay Campbell (finlaycampbell93@gmail.com)

Examples

if (require(outbreaks)) {

## example using MERS outbreak in Korea, 2014
head(mers_korea_2015[[1]])
head(mers_korea_2015[[2]])

x <- make_epicontacts(linelist = mers_korea_2015[[1]],
                       contacts = mers_korea_2015[[2]],
                       directed = TRUE)

if (FALSE) {
plot(x,
x_axis = 'dt_onset',
network_shape = 'rectangle',
node_size = 'R_i',
edge_label = 'exposure',
node_colour = 'sex',
edge_color = 'diff_dt_onset',
node_color = 'sex',
size_range = c(5, 30),
height = 2500,
selector = FALSE,
label = FALSE,
axis_type = 'double',
date_labels = "%b %d",
width = 1500)
}
}

#> Warning: Cycle(s) detected in the contact network: this may be unwanted